requirements.txt
sounddevice==0.4.6
numpy==1.24.3
scipy==1.10.1
openai-whisper==20231117
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
gTTS==2.3.2
python-multipart==0.0.6
websockets==12.0
streamlit==1.32.0 